+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 - Fixed incorrect transaction event with respect to a transaction status for when a delegator decreased its stake.
 - Added support for GRPC V2 `GetBlockInfo` for retrieving information of the block being queried.
 - Added support for GRPC V2 `GetTokenomicsInfo` for acquiring various tokenomics related information for a block.
-- Added support for GRPC V2 `GetInstanceList`
+- Added support for GRPC V2 `GetInstanceList` for retrieving the contract instances at a certain block.
 
 ## 4.2.0
 - Added initial support for GRPC V2